utils/requirements.txt

pymongo
SQLAlchemy
elasticsearch>=8.11.0
redis
kafka-python
pika
websockets
fastapi
uvicorn
requests
PyJWT
cryptography
pydantic>=2
schedule
watchdog
boto3
pandas
PyYAML
bcrypt
pyspark
